moved CSD file methods to csd_file.py
