+++ ecris/csd/viewer/files/__init__.py
@@ -1,0 +1,1 @@
+from .csd_file import CSDFile
