+++ ecris/csd/viewer/files/file_list.py
@@ -1,25 +1,8 @@
-import datetime as dt
-from dataclasses import dataclass
 from pathlib import Path
 import tkinter as tk
 from typing import List
 
-class CSDFile():
-    def __init__(self, path):
-        self.path = path
-        self.datetime_format: str = "%Y-%m-%d %H:%M:%S"
-        self.plotted: bool = False
-        self.file_size: float = 0
-
-    @property
-    def formatted_datetime(self) -> str:
-        time_stamp = self.path.name[-10:]
-        print(time_stamp)
-        return dt.datetime.fromtimestamp(float(time_stamp)).strftime(self.datetime_format)
-
-    @property
-    def list_value(self) -> str:
-        return f"{self.formatted_datetime} ({self.path.name})"
+from .csd_file import CSDFile, get_files
 
 class FileList(tk.Frame):
     def __init__(self, path: Path, *args, **kwargs):
@@ -41,20 +24,13 @@ class FileList(tk.Frame):
     def update_label(self):
         self.directory_label.config(text=f"Current directory: {self.current_directory}")
     
-    def get_files(self):
-        self.files = [CSDFile(p) for p in reversed(
-            sorted(Path(self.current_directory).glob("csd_*")))]
-
     def get_selected_filename(self) -> Path:
         for i in self.file_listbox.curselection():
-            return Path(self.current_directory) / self.file_listbox.get(i)
-    
-    def get_most_recent_filename(self) -> Path:
-        return Path(self.current_directory) / self.file_listbox.get(0)
+            return self.files[i].path
 
     def populate_listbox(self):
         """Populates the listbox with files from the specified directory."""
-        self.get_files()
+        self.files = get_files(self.current_directory)
         self.file_listbox.delete(0, tk.END)
         if not self.files:
             self.stringvar.set(["No CSD files found"])
